fix:过滤空table

+++ tools/goctl/rpc/generator/curdGen.go
@@ -32,7 +32,7 @@ func (l *{{.logicName}}) {{.method}} (in {{.request}}) ({{.response}}, error) {
 		logx.Error(err)
 	}
 	
-	return &{{.responseType}}{ {{.pK}} :id }, nil
+	return &{{.responseType}}{ {{.pK}} :uint64(id) }, nil
 }
 `
